3 - refactor generics to mod / separate dyn to own sub-mod
